[Bugfix] [Enhancement] 热更新融入项目，修复LocalProject
