+++ mosaic-starter/src/main/java/io/github/tml/mosaic/config/MosaicInitConfig.java
@@ -24,7 +24,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
 
-import java.nio.charset.CoderMalfunctionError;
 import java.util.ArrayList;
 import java.util.List;
