+++ mosaic-test-plugin/src/main/java/io/github/tml/plugin/systemLog/api/PropertySystemLogCubeApi.java
@@ -1,23 +1,22 @@
-package io.github.tml.systemLog.api;
+package io.github.tml.plugin.systemLog.api;
 
 import io.github.tml.mosaic.core.tools.guid.GUID;
 import io.github.tml.mosaic.cube.ExtensionPackage;
 import io.github.tml.mosaic.cube.MExtension;
 import io.github.tml.mosaic.cube.MExtensionPackage;
-import io.github.tml.systemLog.cube.SystemLogCube;
+import io.github.tml.plugin.systemLog.cube.PropertySystemLogCube;
 
-import static io.github.tml.systemLog.config.Constant.PLUGIN_ID;
-
+import static io.github.tml.plugin.systemLog.config.Constant.PLUGIN_ID_P;
 
 @MExtensionPackage(value = "system.log",
         name = "系统日志输出扩展包",
         description = "用于系统日志输出",
         version = "1.0.1",
-        cubeId = PLUGIN_ID)
-public class SystemLogCubeApi extends ExtensionPackage<SystemLogCube> {
+        cubeId = PLUGIN_ID_P)
+public class PropertySystemLogCubeApi extends ExtensionPackage<PropertySystemLogCube> {
 
 
-    public SystemLogCubeApi(SystemLogCube cube, GUID guid) {
+    public PropertySystemLogCubeApi(PropertySystemLogCube cube, GUID guid) {
         super(cube, guid);
     }
