+++ mosaic-user-test-project/src/main/java/io/github/tml/plugin/systemLog/api/SystemLogCubeApi.java
@@ -4,20 +4,20 @@ import io.github.tml.mosaic.core.tools.guid.GUID;
 import io.github.tml.mosaic.cube.ExtensionPackage;
 import io.github.tml.mosaic.cube.MExtension;
 import io.github.tml.mosaic.cube.MExtensionPackage;
-import io.github.tml.plugin.systemLog.cube.PropertySystemLogCube;
+import io.github.tml.plugin.systemLog.cube.SystemLogCube;
 
-import static io.github.tml.plugin.systemLog.config.Constant.PLUGIN_ID_P;
-import static io.github.tml.plugin.systemLog.config.Constant.PLUGIN_ID_S;
+import static io.github.tml.plugin.systemLog.config.Constant.PLUGIN_ID;
+
 
 @MExtensionPackage(value = "system.log",
         name = "系统日志输出扩展包",
         description = "用于系统日志输出",
         version = "1.0.1",
-        cubeId = PLUGIN_ID_P)
-public class SystemLogCubeApi extends ExtensionPackage<PropertySystemLogCube> {
+        cubeId = PLUGIN_ID)
+public class SystemLogCubeApi extends ExtensionPackage<SystemLogCube> {
 
 
-    public SystemLogCubeApi(PropertySystemLogCube cube, GUID guid) {
+    public SystemLogCubeApi(SystemLogCube cube, GUID guid) {
         super(cube, guid);
     }
