+++ mosaic-user-test-project/src/main/java/io/github/tml/plugin/systemLog/config/Constant.java
@@ -1,4 +1,4 @@
-package io.github.tml.systemLog.config;
+package io.github.tml.plugin.systemLog.config;
 
 public class Constant {
